feat: convert core project to esm and use vite for umd bundling browser ready version
