+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules/
 /out/
-src/main/.DS_Store
+**/.DS_Store
 .yarn/**
 **.gz
