+++ index.html
@@ -1,0 +1,1 @@
+<p>Hello Vite!</p>
