+++ package.json
@@ -3,6 +3,7 @@
   "version": "4.15.0",
   "main": "index.js",
   "types": "index.d.ts",
+  "type": "module",
   "engines": {
     "node": "^20 || ^22 || ^23"
   },
@@ -14,17 +15,16 @@
     "prepare": "husky",
     "precommit": "lint-staged",
     "lint": "eslint {src,tests}",
-    "rollup": "rollup --config",
-    "uglify": "uglifyjs dist/index.js --compress -o dist/index.min.js",
-    "dist": "npm run rollup && npm run uglify",
     "tsc:postversion": "cp package.json .. && cp package-lock.json ..",
     "swc:build": "swc src --out-dir out",
-    "swc:build:link": "npm run prepack && npm run cp:files && npm link ./out/src",
+    "swc:build:link": "npm run swc:build && npm run swc:cp:types && npm run swc:cp:files && npm link ./out/src",
     "swc:gen:types": "tsc -p tsconfig.types.json --declaration --emitDeclarationOnly",
     "swc:cp:types": "npm run swc:gen:types && cp -R ./types/* ./out/src",
     "swc:cp:files": "cp ./readme.md ./package*.json ./out/src && cp -R ./media ./out/src",
     "swc:sim:pack": "npm run swc:build && npm run swc:cp:files && npm run swc:cp:types",
-    "postversion": "cp package.json ../.. && cp package-lock.json ../.."
+    "postversion": "cp package.json ../.. && cp package-lock.json ../..",
+    "clean": "rm -rf out dist types",
+    "vite:dist": "npm run swc:build && vite build"
   },
   "keywords": [],
   "license": "MIT",
@@ -35,31 +35,24 @@
   },
   "devDependencies": {
     "@jest/types": "^29.6.3",
-    "@rollup/plugin-commonjs": "^28.0.2",
-    "@rollup/plugin-node-resolve": "^16.0.0",
-    "@rollup/plugin-typescript": "^12.1.2",
     "@swc/cli": "^0.6.0",
-    "@swc/core": "^1.10.9",
+    "@swc/core": "^1.10.12",
     "@swc/jest": "^0.2.37",
     "@types/jest": "^29.5.14",
-    "@types/node": "^22.10.10",
-    "@types/path-browserify": "^1.0.3",
+    "@types/node": "^22.13.0",
     "@types/semantic-release": "^20.0.6",
     "eslint": "^9.19.0",
     "eslint-plugin-jest": "^28.11.0",
     "husky": "^9.1.7",
     "jest": "^29.7.0",
-    "lint-staged": "^15.4.2",
+    "lint-staged": "^15.4.3",
     "prettier": "^3.4.2",
-    "rollup": "^4.32.0",
-    "rollup-plugin-polyfill-node": "^0.13.0",
-    "ts-node": "^10.9.2",
     "typescript": "^5.7.3",
-    "typescript-eslint": "^8.21.0",
-    "uglify-js": "^3.19.3"
+    "typescript-eslint": "^8.22.0",
+    "vite": "^6.0.11",
+    "vite-plugin-node-polyfills": "^0.23.0"
   },
   "dependencies": {
-    "path-browserify": "^1.0.1",
     "xmlbuilder2": "^3.1.1"
   }
 }
