+++ src/main/core.ts
@@ -1,0 +1,38 @@
+import { Flow } from "./models/Flow";
+import { FlowAttribute } from "./models/FlowAttribute";
+import { FlowElement } from "./models/FlowElement";
+import { FlowNode } from "./models/FlowNode";
+import { FlowResource } from "./models/FlowResource";
+import { ResultDetails } from "./models/ResultDetails";
+import { RuleResult } from "./models/RuleResult";
+import { ScanResult } from "./models/ScanResult";
+import { FlowType } from "./models/FlowType";
+import { FlowVariable } from "./models/FlowVariable";
+
+import { Compiler } from "./libs/Compiler";
+import { fix } from "./libs/FixFlows";
+import { getRules } from "./libs/GetRuleDefinitions";
+import { parse } from "./libs/ParseFlows";
+import { scan } from "./libs/ScanFlows";
+
+import type { IRulesConfig } from "./interfaces/IRulesConfig";
+import type { IRuleDefinition } from "./interfaces/IRuleDefinition";
+
+export {
+  Flow,
+  FlowAttribute,
+  FlowElement,
+  FlowNode,
+  FlowResource,
+  FlowType,
+  FlowVariable,
+  ScanResult,
+  RuleResult,
+  ResultDetails,
+  Compiler,
+  getRules,
+  parse,
+  scan,
+  fix,
+};
+export type { IRuleDefinition, IRulesConfig };
