+++ tsconfig.types.json
@@ -19,6 +19,7 @@
     ".idea/**",
     "src/data/CustomRuleExample.ts",
     "tests/**/*.test.ts",
-    "jest.config.ts"
+    "jest.config.ts",
+    "vite.config.ts"
   ],
 }
