--- tsconfig.umd.json
@@ -1,27 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "ESNext",
-    "target": "es6",
-    "outDir": "dist",
-    "moduleResolution":"node",
-    "lib": [
-      "es2020.string",
-      "es6"
-    ],
-    "experimentalDecorators": true,
-    "sourceMap": false,
-    "rootDir": "src",
-    "strict": false, /* enable all strict type-checking options */
-    "baseUrl": ".",
-    "declaration": true,
-    "esModuleInterop": true
-  },
-  "exclude": [
-    "node_modules",
-    "out/**",
-    ".idea/**",
-    "src/data/CustomRuleExample.ts",
-    "tests/**/*.test.ts",
-    "jest.config.ts"
-  ]
-}
